Update release notes workflow (#415)
## What is the goal of this PR?

We integrate the new release notes tooling. The release notes are now to be written by a person and committed to the repo.

## What are the changes implemented in this PR?

`tool/release/create_notes.sh` is a manual replacement for automatic `@vaticle_dependencies//tool/release/notes:create` in CI. The script is meant to be executed manually by the writer of the release notes, who will edit the autocompiled notes before committing RELEASE_NOTES_LATEST.md to the repository.

In the Release job in Factory, we check to see whether there were any changes committed since the release notes have last been changed, as a safeguard against accidentally releasing with stale release notes.