Concept API (#43)
## What is the goal of this PR?

We implement the Concept data structures and their corresponding API
requests to the database.

## What are the changes implemented in this PR?

Concept API is implemented as `transaction::concept` extensions to the
passive Concept data structures. We provide an `..API` extension trait
for each Concept type that can be imported as required. This was done in
order to break the circular dependency between the `concept` and
`transaction` modules, clearly separating the data and behaviour.

BDD steps have been refactored to utilize Cucumber's `Parameter` trait
that enable automatic parsing of the steps contents to native types. We
establish a convention by which the parameters can encapsulate the
details that do not affect the rest of the step, e.g. `contain` / `do
not contain` assertions.

As a consequence, we were able to deduplicate steps with optional
arguments, e.g. `owns attribute` / `owns explicit attribute` / `owns
attribute as attribute` can all be handled as
`owns{optional_explicit}{optional_override_label}`. This allowed us to
greatly reduce the amount of duplicated code in the BDD test steps
implementation for Concept API.

Closes #8.