Fix filenames with incorrect symbols for Windows (#500)
## What is the goal of this PR?

We had some filenames in generated AsciiDoc files that are prohibited in Windows. It affected only NodeJS Driver, but cloning the repo was impossible in Windows. We fixed it.

## What are the changes implemented in this PR?

We replaced symbols `"<"` and `">"` with underscores.

We also fixed the script for docs testing: now we add all new generated files to git.