Wrap native exceptions with python driver exception (#498)
## What is the goal of this PR?

We had two exception classes: `TypeDBDriverException`, that could be raised by native calls, and `TypeDBDriverExceptionExt(TypeDBDriverException)`, that was raised by python driver methods. In order to catch all driver exceptions we had to use `TypeDBDriverException`, but this class couldn't be commented properly, because it is defined in the SWIG-generated file, that we wouldn't like to expose.

Now all raised exceptions are `TypeDBDriverException`s, and this class is properly documented.

## What are the changes implemented in this PR?

We catch all native TypeDB exceptions, that could be thrown by native functions calls, and raise `TypeDBDriverException` (formerly known as `TypeDBDriverExceptionExt`).