Directly generate docs into directories and add missing doc comments (#503)
## What is the goal of this PR?

We had all generated docs files in one directory for each driver. It was
not convenient for the user and for the following processing. We added
files containing mapping from file names to directory names and organise
files according to this mapping.

## What are the changes implemented in this PR?

Each driver's directory contains a file `doc_structure.bzl` with a
mapping `dir_mapping` from file names to directory names. This structure
is passed to the `docs_adoc` rule as `docs_dirs` argument. If at least
one file is not mentioned in the file, `NullPointerException` is thrown.

Minor changes:
- We made all methods' anchors unique throughout the driver
documentation.
- Methods, fields and enum constants are ordered alphabetical.
- We added `Label` and `TypeDBDriverException` classes to the Java
Driver docs.