Node driver package automatic version stamping (#517)
## What is the goal of this PR?

We leverage Bazel's built-in workspace status and stamping capabilities
to ensure that the version of TypeDB Protocol depended on by the node
package doesn't go out of sync with the bazel dependency. To that end,
we also add a snapshot deployment test for the node driver.

## What are the changes implemented in this PR?

Drive-by fix: handshake with core server no longer overrides the
user-provided address.