Optimise CI times by retaining server between Java BDD scenarios (#521)
## What is the goal of this PR?

We optimise Java CI time by not shutting down the TypeDB server between
scenarios. Instead, we delete the existing databases each test, which is
much faster.

During this work, we also discovered some sub-par UX in terms of error
messages thrown, and missing BDD steps that needed to be implemented.

## What are the changes implemented in this PR?

* Only boot up TypeDB during the Java BDD Feature setup
* Shut down TypeDB after the last behaviour scenario
* Ignore tests using server configuration - we should only run these
from the backend.
* Clean up various Rust and Node error messsages
* Fix `ignore` tags that we parse from BDD
* This led to new scenarios running that we were missing BDD
implementation in Rust.