Java JNI library loading: fallback when platform not specified (#535)
## Usage and product changes

Previously, the JNI library would be selected based on if the containing
JAR contains the expected platform string. When TypeDB Driver Java is
repackaged by the end user, the JNI library is relocated and is likely
missing the platform specification in its path. Now, if only one native
library candidate is found in classpath, we attempt to use that rather
than fail.