Sync dependencies in CI (#571)
## Usage and product changes

We add a sync-dependencies job to be run in CI after successful snapshot
and release deployments. The job sends a request to vaticle-bot to
update all downstream dependencies.

The snapshot dependencies sync job is run in Factory CI after all
behaviour and integration tests pass. The CircleCI tests only verify
deployment, which is not extensive enough verification to trigger
downstream propagation, and not strictly necessary for sync as that is
only affected by git dependencies.

The release dependencies sync job, in contrast, runs in CircleCI and
only runs after all deployments have succeeded and been verified, so
that all downstream deployments can be safely updated.

Importantly, due to the way sync-dependencies is implemented, we revert
`dependencies/vaticle/repositories.bzl` from having a dedicated
typedb-protocol version line for the purposes of workspace-status, back
to inlined version with the sync-marker. This means that the sync-marker
is performing double duty as a workspace status marker.

Note: this PR does _not_ update the `dependencies` repo dependency. It
will be updated automatically by the bot during its first pass.