Setup python3.11 CFFI in CI for sync-dependencies (#579)
## Usage and product changes

CI executors based on Ubuntu 22.04 default to python 3.10. Due to us
having to be explicit about which version of Python we are using for
native addon compilation for FFI driver, we explicitly install
python3.11, pip, and the cffi package before starting a
sync-dependencies job. This eliminates the potential for mismatch in the
future when defaults change.