Fix parsing & formatting issues for C, C++ documentation (#619)
## Usage and product changes
Fixes the automatic documentation generation tools for the C & C++
drivers (1) to be consistent with other languages in terms of structure,
and (2) removes stray duplication of parameter documentation in method
descriptions.

## Implementation
* Introduces artificial headers to the C driver based on filename, since
methods are logically grouped into files. In other drivers, the grouping
is done by class.
* Fixes the C & C++ parsers to not duplicate parameters in method
description when examples are absent.