Fix misleading cloud encryption error message (#645)
## Usage and product changes
The old version of the cloud encryption error message confused the user
in case, for example, their endpoint is not encrypted, but the
connection is. There are also other potential causes of the `received
corrupt message` status message, that we can't understand on a deeper
level, so it's more correct to have a less specific error message here.

## Implementation
Renamed the error.