Fix aspect_bazel_lib dependency, update typedb and nodejs dependencies (#658)
## Usage and product changes
1. We update `typedb` dependencies to `2.28.3` before the release to
verify that the builds are still correct before the `driver` release.
2. We update other dependencies that have previously broken the build
because of the updated `ts` (`nodejs`) rules:
**dependencies** ref with `nodejs` and `csharp` rules becoming more
mature based on [aspect_bazel_lib version
globalization](https://github.com/vaticle/dependencies/pull/547);
**typedb-protocol** ref to
[propagate](https://github.com/vaticle/typedb-protocol/pull/202) the
fixed `ts` rules and fix its build based on this propagation.

## Implementation
We use the updated rules fixed in
[dependencies](https://github.com/vaticle/dependencies/pull/547) and
also update the code based on the new requirements of the updates `ts`
rules. More details in the comments.