Remove excessive diagnostics flags, turn on development mode for tests (#659)
## Usage and product changes
We clean the usage of `--diagnostics.***` flags used in the test servers
because of the recent propagation of a newer version of `TypeDB`, now
supporting the `--development-mode.enable` flag. As the development mode
includes disabling any outer reporting, it is a perfect substitution for
this logic.

## Implementation
We remove the unnecessary flags everywhere, aiming to cover it by the
activated development mode.

* This mode is activated by default for all the `TypeDB***Runner`s
[relevant for the `Java` driver]
* This mode is **not** activated for `vaticle_typedb_artifact` fetched
by `vaticle_typedb_artifact()` or built by `bazel build
@vaticle_typedb_artifact_*platform*//file`, so we need to manually add
`--development-mode.enable=true` for all the invocations of these
artifacts.