3.0 Java driver (#684)
## Usage and product changes
We introduce the updated Java driver for the upcoming 3.0 release. To do
so, we update the C APIs to align with the updated Rust driver, add the
necessary changes to the SWIG layer and reimplement the Java driver
itself.

It is already used in integration tests (not in the CI) and in [the
updated console](https://github.com/typedb/typedb-console/pull/258).

## Implementation
Non-existing features are deleted. Non-implemented features are
commented out (cloud, user management, options).

The release pipeline should be the same. However, we may need to update
the dependencies on typedb artifacts later in the following smaller prs.

No readme updates for now.