Introduce 3.0 Python driver (#692)
## Usage and product changes
We introduce the updated Python driver for the upcoming 3.0 release. To
align with the updated Rust driver, we removed `Concept API` (so you
could simplify your querying workflow with a single `tx.query()` entry
point), squeezed sessions and transactions to standalone transactions,
and remodeled messaging with the server.

As it's an alpha release, some of the features are temporarily disabled:
* Options;
* User management;
* Cloud connection with replicas information.

Moreover, we no longer support Python 3.8 as [its support comes to an
end](https://devguide.python.org/versions/) and we want to offer the
full support of our timezones with the standard library equally for all
the versions of the language, while `ZoneInfo` from the standard library
has been introduced in version 3.9.

## Implementation
Non-existing features are deleted. Non-implemented features (mentioned
in the previous section) are commented out.

The build is returned to the CI, leaving only the behaviour and cloud
integration tests commented out.

No readme updates for now.

Some `c` and `java` docs are updated together with the new python docs
(no docs regeneration, just the code).