Introduce 3.0 Python driver docs and tests (#696)
## Usage and product changes
We introduce the second version of the Python driver, fixing a number of
minor bugs from the first version, enhancing existing APIs, and
presenting the updated documentation and automated tests.

Major changes:
* Introduce [3.0 driver BDD
tests](https://github.com/typedb/typedb-behaviour/pull/303);
* Fully return core Python driver to the CI; 
* Refactor and enrich integration tests to cover custom value types and
prepare examples for `README`;
* Prepare semi-automated generation of usage examples for `README` based
on integration tests running in CI to make sure that it's up to date and
in a needed format in the future;
* Update API references and documentation.

## Implementation
Minor changes and details:
* Remove `Label` classes from Java and Python (3.0) drivers to simply
represent labels as strings;
* Fix Rust's driver's database management's deadlock;
* Python BDDs are reimplemented in a similar to the TypeDB server
fashion (using `MayError`, more typed parsing (had been abandoned in one
of the iterations, judging the state of the code), etc.);
* Semi-automated examples generation could be fully automated, but it
would require much more Bazel work to generate the intermediate code
blocks file and pass it as a file to the second rule. It can be done
later if the task receives enough priority. However, it can already be
used as a one-liner (with two `bazel run` commands), and can already be
used in CIs. With just a bit more dirt than it would ideally generate.