Add concept documents for fetch queries (#698)
## Usage and product changes
We add concept documents to support the results of the reintroduced
`fetch` queries.

In Rust, results of `fetch` are streams of `ConceptDocument`s. It is
possible to work with the structured document as a Rust `struct`, but
it's also possible to convert it to a `JSON` document and/or its
`String` representation.

In Java, these results are presented as iterators over custom `JSON`
class instances. A respective `toString` method is available.

In Python, these results are presented as iterators over standard `dict`
instances (able to be printed).

Additionally, we add `QueryType` getters for general `QueryAnswer`s, so
it's possible to check its type without collection.

Usage examples are shown in `README` for the Rust and the Python
drivers. Example integration tests are also available for all 3 drivers.