Introduce 3.0 Python driver docs and tests. Implement new fetch BDD steps in 3.0 drivers (#700)
## Usage and product changes
We introduce updated documentation, usage examples, and automated tests
for the Java driver, to cover all the existing driver's APIs by
sustainable validations.

Additionally, we implement additional BDD steps to check concept
documents in BDDs for other 3.0 drivers: Rust and Python.

As a small bonus, the protocol has been updated, and `Ok` `QueryAnswer`s
receive correct `QueryType`s from the server instead of the client-side
hardcode usage.

## Implementation
For behaviour tests, the already existing architecture is preserved, and
only the steps are rewritten to match the new definitions.
For integration tests, the new approach of the Python driver is chosen,
and all the excessive tests are removed. Now, we only write `example`
tests and tests to cover language-specific values (especially our
implementations and their interfaces).
For docs, the old instruments are used without changes.
For `README`, the `example` integration test parsing instrument is used
as in Rust and Python.