Remove promises resolves in destructors to eliminate redundant exceptions. Cleanup Python exceptions formatting (#713)
## Usage and product changes
We remove the feature of `Promise`s to call `resolve` in destructors in
languages like Java and Python. Now, if a promise is destroyed, it's not
resolved, and, thus, not awaited. This helps the driver to remove the
excessive duplicated errors printing in destructors of promises,
although the error is already returned from a commit operation.

Now, if you just run queries and close the transaction, nothing will be
awaited and persisted. However, if you commit your transaction, all the
ongoing operations on the server side will finish before the actual
commit. This lets you speed up the query execution (in case these
queries don't collide with each other):
```py
for query in queries:
    tx.query(query)
tx.commit()
```
If one of the `queries` contains an error and it's not `resolve`d, it
will be returned from the `commit` call, and no excessive errors will be
printed on resource release.

Detailed examples for each language supported are presented in
`README`s.

Additionally, Python Driver's `TypeDBDriverError` exceptions no longer
show the excessive traceback of its implementation, and only the short
informative version for your executed code is presented.

## Implementation
We update the swig `.i` file to use `_drop` instead of `_resolve` for
all promises (all network operations). Now, we consider that, if not
resolved/awaited, it's just ignored, just like with `await` in Rust.

To ensure that the "delayed" execution for multiple query requests works
correctly, we could use BDDs. However, BDDs would not check excessive
error messages from destructors (and sometimes it's uncaught, like in
Python), so we use the fact that we (and the users) check the output of
the example integration tests, so we combine the goal to showcase such
usage and the goal to test all these error messages in implementing two
simple integration tests steps.