Refactor concepts APIs to quickly access optional instances and values properties (#712)
## Usage and product changes
We generalize the approach to getting concepts properties in TypeDB
Drivers, introducing a set of new APIs for fetching optional values
related to instances of `Concept` classes in Java and Python.

Now, all subclasses of `Concept` have a set of new interfaces starting
with `try` to access IIDs, labels, value types, and values without a
need to cast to a specific `Instance` or `Value`. These methods can be
useful if:
* you have an established workflow with constant queries and always
expect these properties to exist;
* you want to implement a custom handling of cases where the expected
values are missing shorter (without exceptions).

Additionally, value type checks like `is_boolean`/`isBoolean` are also
declared on the top-level `Concept`.

Note that casting is still possible, and its benefits are, as usual:
* static type checking for your programs using TypeDB Driver;
* access to the non-optional `get` interfaces for specific subclasses of
`Concept` like `get_iid`/`getIID` for `Entity` and `Relation`.

Also note that some of the interfaces have changed (e.g., `asBoolean` ->
`getBoolean` for `Attribute` and `Value` to separate value retrieval and
concept casting), but their functioning has not.

## Implementation
On the C driver side, almost all subconcept-specific methods have been
transformed into `concept_`, receiving just a `Concept` instance. The
only exception is `get_type` for `Entity`, `Relation`, and `Attribute`
instances, as it also uses the inner classes of the `Concept`'s enum in
Rust (I think it's simple enough to leave it this way, but I may also
refactor it if you want to).

On the wrapper's users' side (Python and Java), we use `try_get_`
methods in `Concept` implementations and reuse these methods in `get_`
methods in specific subclasses of `Concept` to throw or return the
unwrapped error (sometimes we throw an inner error if it's our mission
to make sure that the requested value exists, sometimes -- continue
throwing user errors for unsafe gets like for `getBoolean`, just like we
used to).