Introduce 3.0 cloud driver tests (#735)
## Usage and product changes
We fix cloud drivers creation interfaces and introduce integration and
behavior tests to cover these public methods.

## Implementation
The integration tests are currently just a copypaste of the core
integration tests with another interface. Not sure if we want to change
it, but it would be smarter to reuse this code differently in the
future.

The behavior tests are actually currently similar to the core tests:
they don't use encryption and they only use one node. We'll need to
update it with more nodes in action when it's ready. Encryption can be
added earlier in a separate PR: there are all the relevant TODOs in the
code.