Introduce optional concepts to Concept Rows. Cleanup driver errors. Automate README examples updates. (#738)
## Usage and product changes
* **Introduce optional concepts to Concept Rows**
`get()` and `get_index()` interfaces of Concept Rows always return
optional Concepts (previously, the Java and Python drivers used to
return non-optional `Concept` instances). If the requested variable name
or index exists in the column names of a row, but the actual value for
this variable is empty, an empty optional value is returned. This is a
natural behavior for optionals (coming to TypeDB soon!) and is already
useful for queries like (where the variable `$empty` won't have values):
```
match not {$empty isa user;}; insert $u isa user, has username "Hi";
```

If the requested variable name or index does not exist in the column
names of a row, an error is returned.

* **Cleanup driver errors**
Query errors are dissolved in the Java and Python drivers. 
All the arguments passed to the external interfaces are validated to be
non-null and of the correct format (e.g., non-negative for
`concept_row.get_index(column_index)`).

* **Automate README examples updates**
Cloud driver usage examples are added to all the available READMEs.
Additionally, these examples are officially available in the repo as
separate files.

The process of README examples updating is automated by unifying all
language updates in a single script. A new CI job is introduced to
verify that the README examples are up to date.

## Implementation