Merge library targets referring to the same crate (#744)
## Usage and product changes

When two library targets with the same crate name and path appear among
the bazel rust targets, we assume them to refer to the same crate, but
with different configuration (e.g., different set of enabled features).

We update the C library targets with a crate name tag to allow the sync
tool to merge them correctly.

## Implementation