Introduce Rust driver token-based authentication (#749)
## Usage and product changes
We update the protocol version and introduce token-based authentication
for all the available drivers. Now, instead of sending usernames and
passwords for authentication purposes, authorization tokens are
implicitly added to every network request to a TypeDB server. It
enhances the authentication speed and security.

These tokens are acquired as a result of driver instantiation and are
renewed automatically. This feature does not require any user-side
changes.

Additionally, as a part of the HTTP client introduction work, we rename
Concept Documents key style from snake_case to camelCase, which is a
common convention for JSONs (it affects only value types: `value_type`
-> `valueType`).

## Implementation
Every network call adds tokens instead of user credentials and is
repeated after a token renewal request if it receives an authentication
error.