Typescript HTTP driver (#767)
## Usage and product changes

Add a relatively slim typescript driver based on our HTTP API, adapted
from code used for this purpose in `typedb-studio`.

This driver is published to `npm` under the name `typedb-driver-http`
through the same process as the old `nodejs` driver.

## Implementation

Used
https://github.com/typedb/typedb-studio/tree/a3e422544b31175906c41bdb805b51c2c3ebaa60/src/framework/typedb-driver
as a base.

Expanded it to cover all endpoints of the HTTP API as per
https://typedb.com/docs/drivers/http/api-reference

Added release to `npm` based on the code used to release the old
`nodejs` driver.

## Testing

We implement steps for the driver behaviour tests, making use of the
existing `@ignore-typedb-http` to skip invalid tests.