Typescript HTTP driver docs generation (#772)
## Usage and product changes

Introduce docs generation for the Typescript HTTP driver, based on the
existing strategy for the NodeJS driver.

We divide the docs into:

- Connection
- Response
- Concept
- Query Structure

within each section, static functions are grouped into a specific
`Static Functions` section.

Additionally, `QueryConstraints`, `QueryVertices` in query structure;
and `DriverParams` in connection are used as sub-categories (that is,
`DriverParams`, `DriverParamsBasic`, and `DriverParamsTranslated` are
all grouped under `DriverParams`).

We also implement functionality for handling:
- type aliases
- indexable properties (e.g. `[varName: string]: Concept`)

## Implementation

Copy across the `NodejsDocsParser` to the `HTTPTSDocsParser`, and modify
some of the HTML queries to function.

Introduce the new `--static-function` and `--nested-class-prefix` CLI
options, which we use to:
- indicate where certain static functions should be grouped
- indicate which classes should have an additional layer of nesting, for
the sake of categorisation

respectively.

Type aliases are added in full to the `description` of a class as a
chunk of code formatting.

Indexable properties are handled as regular properties are, albeit with
specific handling to parse them from the typedoc HTML.