Make the HTTP TypeScript driver dual-module (CJS+ESM) (#776)
## Usage and product changes

The HTTP TypeScript driver is now dual-module, offering both CommonJS
and ES Module support.

## Motivation

In general, it is good practice to make modules intended for Web
consumption to be dual-module. There are workarounds available to use a
CommonJS module in an ES module environment, but they are really ugly.

## Implementation

Previously, we used `tsc` (via `ts_project` in Bazel) to output a list
of `.js` files and their matching `.d.ts` type definitions. Now, we use
`tsup` (via `js_binary` and `js_run_binary` in Bazel) to output the
three essential components of a dual-module setup:

- `index.cjs` - CommonJS runtime entry point
- `index.mjs` - ES module runtime entry point
- `index.d.ts` - TypeScript type definitions when importing ES module

When built outside the Bazel sandbox, it also generates source maps for
debugging.

As a drive-by, we now bundle `LICENSE` into the final NPM package.

### Current limitations

- `npm_package` doesn't seem to be able to accept files in a higher
directory to its own location, so for now we've copied `LICENSE` to
`http-ts/LICENSE`
- `tsup` is supposed to output `index.d.cts` (type defs when importing
CJS module) but doesn't in Bazel for some reason. ES module is our
primary env though
- `node_modules` needs to exist for the `http-ts` `BUILD` targets to
build; therefore, we now run `pnpm i` outside the sandbox in CI when
necessary