Enable TLS by default in Python (#777)
## Usage and product changes

We want to enable a secure-by-default setting in TypeDB Drivers. In Java
and Rust, `DriverOptions` have to be explicitly set, and there are no
defaults. However, Python features a disabled TLS default. While this is
compatible with TypeDB CE, it's an insecure default & not compatible
with TypeDB Cloud without explicitly enabling it.

Instead, we set the default to TLS being __enabled__ in Python. This
means when using an insecure, plaintext connection the user must
explicitly set it, and is more likely to become aware of the plaintext
communication.

## Implementation

- Fix a documentation string in Rust
- Set the default TLS connection in Python to be TLS enabled, instead of
disabled

- Update some forgotten HTTP-TS docs generation and testing targets