Align HTTP driver analyze type names with rust (#821)
## Usage and product changes
Aligns HTTP driver analyze type names with rust. Any code using the driver will break. Major changes are:
* QueryConstraint* is renamed to Constraint*
* QueryVertex* is renamed to ConstraintVertex*

The older names are preserved in `legacy.ts` but not exported to make updating to the new names easier.
TypeDB 3.7.0 introduces a few minor breaking changes in the HTTP API. 
* Expression constraints now return a single variable instead of a list of variables.
* The structure returned with a ConceptRowResponse has the `blocks` field renamed to `conjunctions`
For applications which must operate with both versions before and after 3.7.0, certain types are exported with 'Legacy' as suffix.