Temporarily make https/tls validation more friendly (#832)
## Usage and product changes
Allow using `http` and `https` schemes in connection URLs, introducing a
temporary comparison of the specified scheme and the TLS driver option.
This is a temporary resolution until we decide how to make the GRPC
endpoint safe and easy to understand, yet comfortable for Cloud and
Cluster.

## Implementation
Additionally, introduce more Java interfaces as a part of its build fix.